feat: add LitElement to recommended config
Introducing the `invalid-extends` rule into the best practice config
resulted in a larger breaking change than expected: LitElement repos
fail to lint without introducing an `elementBaseClasses` setting.

Since a large chunk of our consumers are lit users, it makes sense to
add it to the recommended config so it doesn't need configuring
individually.